Mark internal DBus tests as skipped
Make it a bit more obvious in test results that these are skipped. This
is so tests results output actually reflect what's going on, and to
avoid having commented-out code.

Note: Tests are already being skipped, it's just that the call is
commented out rather than skipped, so this does not reduce the scope of
tests or anything alike.