Try to avoid race condition when receiving unix fd parameters
When using txdbus with BlueZ, there is a race condition when calling the org.bluez.GattCharacteristic1.AcquireWrite method. After sending the request, we receive an org.freedesktop.dbus.Properties.PropertiesChanged signal. First, txdbus.protocol.BasicDBusProtocol.fileDescriptorReceived is called and stores the unix fd, Then txdbus.protocol.BasicDBusProtocol.rawDBusMessageReceived method is called for the PropertiesChanged signal and clears the cached unix fds. Then the return values of AcquireWrite are receviced and the fd ends up being None.

To work around this, we hang on to the unix fds until we receive a message that has a unix fd as a parameter.